requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
django-environ>=0.9
celery>=5.2
redis>=4.5
drf-spectacular>=0.26
weasyprint>=58.0
xhtml2pdf>=0.2.8
python-dotenv>=1.0
pytest>=7.0
pytest-django>=4.5
factory-boy>=3.2
pytest-cov>=4.0
black
ruff
mypy
isort
django-filter>=23.1
